refactor: simplify type checking for enum members in extension decorators
